Relocate DictionaryListPresentation to SearchPresentationList
