+++ frontend/app_v2/src/components/DictionaryList/DictionaryListPresentation.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, { useState } from 'react'
 import PropTypes from 'prop-types'
 import { Link } from 'react-router-dom'
 
@@ -7,8 +7,11 @@ import AudioMinimal from 'components/AudioMinimal'
 import useIcon from 'common/useIcon'
 import { makePlural } from 'common/urlHelpers'
 import { getMediaUrl } from 'common/urlHelpers'
-import ActionsMenu from 'components/ActionsMenu'
 import Loading from 'components/Loading'
+import ActionsMenu from 'components/ActionsMenu'
+import Drawer from 'components/Drawer'
+import Word from 'components/Word'
+import Phrase from 'components/Phrase'
 
 /**
  * @summary DictionaryListPresentation
@@ -18,10 +21,55 @@ import Loading from 'components/Loading'
  *
  * @returns {node} jsx markup
  */
-function DictionaryListPresentation({ actions, isLoading, items, moreActions, sitename, wholeDomain }) {
+
+function DictionaryListPresentation({ actions, infiniteScroll, isLoading, items, moreActions, sitename, wholeDomain }) {
+  const [drawerOpen, setDrawerOpen] = useState(false)
+  const [selectedItem, setselectedItem] = useState({})
+  const { isFetchingNextPage, fetchNextPage, hasNextPage, loadButtonLabel, loadButtonRef } = infiniteScroll
+
+  function getDrawerContents() {
+    const { id, type, parentDialect } = selectedItem
+
+    const link = type ? (
+      <Link
+        className="font-medium text-fv-charcoal ml-5 inline-flex items-center float-left"
+        to={`/${parentDialect?.shortUrl ? parentDialect.shortUrl : sitename}/${makePlural(type)}/${id}`}
+      >
+        Go to {type} {useIcon('ForwardArrow', 'fill-current h-10 w-10')}
+      </Link>
+    ) : null
+
+    switch (type) {
+      case 'word':
+        return (
+          <>
+            {link}
+            <Word.Container docId={selectedItem.id} />
+          </>
+        )
+      case 'phrase':
+        return (
+          <>
+            {link}
+            <Phrase.Container docId={selectedItem.id} />
+          </>
+        )
+      default:
+        return null
+    }
+  }
+
+  function handleItemClick(item) {
+    if (item.type === 'word' || item.type === 'phrase') {
+      setselectedItem(item)
+      setDrawerOpen(true)
+    }
+    return
+  }
+
   return (
     <Loading.Container isLoading={isLoading}>
-      {items !== undefined && items?.length > 0 ? (
+      {items?.pages !== undefined && items?.pages?.length > 0 ? (
         <div className="flex flex-col">
           <div className="py-2 align-middle inline-block min-w-full">
             <div className="shadow-md overflow-hidden border-b border-gray-300 sm:rounded-lg">
@@ -30,26 +78,26 @@ function DictionaryListPresentation({ actions, isLoading, items, moreActions, si
                   <tr>
                     <th
                       scope="col"
-                      className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider"
+                      className="px-6 py-3 text-left text-xs font-medium text-fv-charcoal-light uppercase tracking-wider"
                     >
                       Entry
                     </th>
                     <th
                       scope="col"
-                      className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider"
+                      className="px-6 py-3 text-left text-xs font-medium text-fv-charcoal-light uppercase tracking-wider"
                     >
                       Translation
                     </th>
                     <th
                       scope="col"
-                      className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider"
+                      className="px-6 py-3 text-left text-xs font-medium text-fv-charcoal-light uppercase tracking-wider"
                     >
                       Type
                     </th>
                     {wholeDomain ? (
                       <th
                         scope="col"
-                        className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider"
+                        className="px-6 py-3 text-left text-xs font-medium text-fv-charcoal-light uppercase tracking-wider"
                       >
                         Language Site
                       </th>
@@ -60,104 +108,110 @@ function DictionaryListPresentation({ actions, isLoading, items, moreActions, si
                   </tr>
                 </thead>
                 <tbody className="bg-white divide-y divide-gray-300">
-                  {items.map(({ id, title, translations, audio, type, parentDialect }, index) => (
-                    <tr key={id + index}>
-                      <td className="px-6 py-4 flex items-center">
-                        <Link
-                          className="font-medium text-gray-900 mr-2"
-                          to={`/${parentDialect?.shortUrl ? parentDialect.shortUrl : sitename}/${makePlural(
-                            type
-                          )}/${id}`}
-                        >
-                          {title}
-                        </Link>
-                        {audio.length > 0
-                          ? audio.map((audioId, i) => (
-                              <AudioMinimal.Container
-                                key={i}
-                                src={getMediaUrl({ id: audioId, type: 'audio' })}
-                                icons={{
-                                  Play: useIcon(
-                                    'Audio',
-                                    'fill-current text-fv-charcoal-light hover:text-fv-charcoal m-1 h-8 w-8 md:h-6 md:w-6'
-                                  ),
-                                  Pause: useIcon(
-                                    'PauseCircle',
-                                    'fill-current text-fv-charcoal-light hover:text-fv-charcoal m-1 h-8 w-8 md:h-6 md:w-6'
-                                  ),
-                                  Error: useIcon(
-                                    'TimesCircle',
-                                    'fill-current text-fv-charcoal-light hover:text-fv-charcoal m-1 h-8 w-8 md:h-6 md:w-6'
-                                  ),
-                                }}
-                              />
-                            ))
-                          : null}
-                      </td>
-                      <td className="px-6 py-4">
-                        {translations ? (
-                          <ol className="text-gray-900">
-                            {translations.map((translation, i) => (
-                              <li key={i}>
-                                {translations.length > 1 ? `${i + 1}. ` : null} {translation}
-                              </li>
-                            ))}
-                          </ol>
-                        ) : null}
-                      </td>
-                      <td className="px-6 py-4 whitespace-nowrap">
-                        <span
-                          className={`px-2 inline-flex text-xs leading-5 font-semibold rounded-full bg-${type} capitalize text-white`}
-                        >
-                          {type}
-                        </span>
-                      </td>
-                      {wholeDomain ? (
-                        <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
-                          <Link to={`/${parentDialect.shortUrl}`}>{parentDialect.name}</Link>
-                        </td>
-                      ) : null}
-                      <td className="text-right px-6">
-                        <ActionsMenu.Container
-                          docId={id}
-                          docTitle={title}
-                          docType={type}
-                          actions={actions}
-                          moreActions={moreActions}
-                          withConfirmation
-                          withTooltip
-                        />
-                      </td>
-                    </tr>
+                  {items.pages.map((page, index) => (
+                    <React.Fragment key={index}>
+                      {page.results.map(({ id, title, translations, audio, type, parentDialect }) => (
+                        <tr key={id}>
+                          <td className="px-6 py-4 flex items-center">
+                            <button
+                              className="font-medium text-fv-charcoal mr-2"
+                              onClick={() => handleItemClick({ id, type, parentDialect })}
+                            >
+                              {title}
+                            </button>
+                            {audio.length > 0
+                              ? audio.map((audioId, i) => (
+                                  <AudioMinimal.Container
+                                    key={i}
+                                    src={getMediaUrl({ id: audioId, type: 'audio' })}
+                                    icons={{
+                                      Play: useIcon(
+                                        'Audio',
+                                        'fill-current text-fv-charcoal-light hover:text-fv-charcoal m-1 h-8 w-8 md:h-6 md:w-6'
+                                      ),
+                                      Pause: useIcon(
+                                        'PauseCircle',
+                                        'fill-current text-fv-charcoal-light hover:text-fv-charcoal m-1 h-8 w-8 md:h-6 md:w-6'
+                                      ),
+                                      Error: useIcon(
+                                        'TimesCircle',
+                                        'fill-current text-fv-charcoal-light hover:text-fv-charcoal m-1 h-8 w-8 md:h-6 md:w-6'
+                                      ),
+                                    }}
+                                  />
+                                ))
+                              : null}
+                          </td>
+                          <td className="px-6 py-4">
+                            {translations ? (
+                              <ol className="text-fv-charcoal">
+                                {translations.map((translation, i) => (
+                                  <li key={i}>
+                                    {translations.length > 1 ? `${i + 1}. ` : null} {translation}
+                                  </li>
+                                ))}
+                              </ol>
+                            ) : null}
+                          </td>
+                          <td className="px-6 py-4 whitespace-nowrap">
+                            <span
+                              className={`px-2 inline-flex text-xs leading-5 font-semibold rounded-full bg-${type} capitalize text-white`}
+                            >
+                              {type}
+                            </span>
+                          </td>
+                          {wholeDomain ? (
+                            <td className="px-6 py-4 whitespace-nowrap text-sm text-fv-charcoal-light">
+                              <Link to={`/${parentDialect.shortUrl}`}>{parentDialect.name}</Link>
+                            </td>
+                          ) : null}
+                          <td className="text-right px-6">
+                            <ActionsMenu.Container
+                              docId={id}
+                              docTitle={title}
+                              docType={type}
+                              actions={actions}
+                              moreActions={moreActions}
+                              withConfirmation
+                              withTooltip
+                            />
+                          </td>
+                        </tr>
+                      ))}
+                    </React.Fragment>
                   ))}
                 </tbody>
               </table>
+            </div>
+            <div>
+              <button
+                ref={loadButtonRef}
+                className={'px-2 inline-flex text-xs leading-5 font-semibold rounded-full bg-primary text-white'}
+                onClick={() => fetchNextPage()}
+                disabled={!hasNextPage || isFetchingNextPage}
+              >
+                {loadButtonLabel}
+              </button>
             </div>
           </div>
         </div>
       ) : (
         <div className="m-5 md:mx-20 md:my-10">Sorry, no results were found for this search.</div>
       )}
+      <Drawer.Presentation isOpen={drawerOpen} closeHandler={() => setDrawerOpen(false)}>
+        {getDrawerContents()}
+      </Drawer.Presentation>
     </Loading.Container>
   )
 }
 
 // PROPTYPES
-const { array, arrayOf, bool, shape, string } = PropTypes
+const { array, bool, object, string } = PropTypes
 DictionaryListPresentation.propTypes = {
   actions: array,
+  infiniteScroll: object,
   isLoading: bool,
-  items: arrayOf(
-    shape({
-      id: string,
-      title: string,
-      path: string,
-      translations: array,
-      audio: array,
-      type: string,
-      sitename: string,
-    })
-  ),
+  items: object,
   moreActions: array,
   sitename: string,
   wholeDomain: bool,
